refactor: Symptom 상수 data 분리
