+++ client/src/features/firstAid/data/symptom.ts
@@ -1,0 +1,13 @@
+export const SYMPTOMS = [
+  { code: 'BLEEDING', label: 'BLEEDING' },
+  { code: 'BURNS', label: 'BURNS' },
+  { code: 'FRACTURE', label: 'FRACTURE' },
+  { code: 'ALLERGIC_REACTION', label: 'ALLERGIC REACTION' },
+  { code: 'SEIZURE', label: 'SEIZURE' },
+  { code: 'HEATSTROKE', label: 'HEATSTROKE' },
+  { code: 'HYPOTHERMIA', label: 'HYPOTHERMIA' },
+  { code: 'POISONING', label: 'POISONING' },
+  { code: 'BREATHING_DIFFICULTY', label: 'BREATHING DIFFICULTY' },
+  { code: 'ANIMAL_BITE', label: 'ANIMAL_BITE' },
+  { code: 'FALL_INJURY', label: 'FALL INJURY' },
+];
