+++ client/src/widgets/bottomNavigateBar/data/BottomNavigateBarData.ts
@@ -3,17 +3,3 @@ export const MENU_ITEMS = [
   { code: 'EMERGENCY', label: 'Emergency Call' },
   { code: 'CONTACT', label: 'Contact Us' },
 ];
-
-export const SYMPTOMS = [
-  { code: 'BLEEDING', label: 'BLEEDING' },
-  { code: 'BURNS', label: 'BURNS' },
-  { code: 'FRACTURE', label: 'FRACTURE' },
-  { code: 'ALLERGIC_REACTION', label: 'ALLERGIC REACTION' },
-  { code: 'SEIZURE', label: 'SEIZURE' },
-  { code: 'HEATSTROKE', label: 'HEATSTROKE' },
-  { code: 'HYPOTHERMIA', label: 'HYPOTHERMIA' },
-  { code: 'POISONING', label: 'POISONING' },
-  { code: 'BREATHING_DIFFICULTY', label: 'BREATHING DIFFICULTY' },
-  { code: 'ANIMAL_BITE', label: 'ANIMAL_BITE' },
-  { code: 'FALL_INJURY', label: 'FALL INJURY' },
-];
